Prometheus ec2 monitoring (#182)
Originally this PR was about adding prometheus monitoring to services
and nodes in the ECS cluster, but then we realized that the application
level metrics are trickier to implement because ECS deployed tasks have
a random port assigned, but [the standard ec2 discovery settings in
Prometheus](https://prometheus.io/docs/prometheus/latest/configuration/configuration/#ec2_sd_config)
require you to provide the port in advance. So for this reason in this
PR we only set up node level metrics, since nodes can have a node
exporter process running in a fixed port (see #179).

Since the ECS nodes are not reachable through the Internet, we added a
proxy server to forward scrape requests from the monitoring server to
the actual nodes.

To achieve node level metrics scraping, we: 

- Add an Nginx proxy configuration to the clickhouse proxy server to
direct traffic from the monitoring server to the actual nodes. In order
to know to which node is directed each scraping request, we send the
private IP address of the node as a path parameter and using nginx rules
we parse the right host to send the metrics request to
- Add a Prometheus relabeling configuration that will take what the
standard ec2 discovery settings provide and use it to rewrite the
address to point to the proxy server and adding the private IP as a
parameter
- Add the permission configurations required to allow traffic from the
monitoring server to the proxy server, and from the proxy server to the
cluster nodes

 This PR solves #171 and #172 and depends on #179