Add tag filtering monitoring (#217)
Sometimes we don't want the monitoring system to monitor all machines.
In some cases a machine doesn't need to be monitored but it will, and if
it's not properly set it will error and show up as down in Prometheus,
causing noise. For example: The Ansible controller machine or the
codesign box

In this PR I add a new tag `MonitoringActive`  to EC2 nodes used in: 
- The EC2 module 
- The ecs_cluster module
- The HSM module
- The ansible controller module

So that Prometheus can focus on monitoring machines with the tag set to
true: `MonitoringActive = 'true'` and ignore the rest

closes #215