Add prometheus exporter to fastpath
note that I use network mode = host for the fastpath to use docker to
match the ports of statsd expected by the fastpath and
the one exposed by the exporter.

Both the exporter and the fastpath have hardcoded ports
that cannot be changed with configuration. So this is the easiest
way to forward traffic from fastpath to the exporter